[Modified] MusicContent 인터페이스 수정
